+++ core/equipment.c
@@ -14,7 +14,6 @@
 #include "equipment.h"
 #include "gettext.h"
 #include "dive.h"
-#include "display.h"
 #include "divelist.h"
 #include "pref.h"
 #include "subsurface-string.h"
