+++ core/gaspressures.c
@@ -21,7 +21,6 @@
 #include "ssrf.h"
 #include "dive.h"
 #include "event.h"
-#include "display.h"
 #include "profile.h"
 #include "gaspressures.h"
 #include "pref.h"
