+++ qt-models/filtermodels.cpp
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: GPL-2.0
 #include "qt-models/filtermodels.h"
-#include "core/display.h"
 #include "core/qthelper.h"
 #include "core/trip.h"
 #include "core/subsurface-string.h"
